+++ src/main/java/coupon/config/ReadOnlyDataSourceRouter.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.config;
 
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
