+++ src/main/java/coupon/domain/Member.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
