+++ src/main/java/coupon/repository/CouponRepository.java
@@ -1,7 +1,9 @@
-package coupon;
+package coupon.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import coupon.domain.Coupon;
 
 @Repository
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
